Avance 05/06/2024

+++ semana12/store_api/app/models/producto_model.py
@@ -1,4 +1,4 @@
-from database import db
+from app.database import db
 
 
 # Define la clase `Animal` que hereda de `db.Model`
@@ -10,7 +10,7 @@ class Producto(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False)
     description = db.Column(db.String(100), nullable=False)
-    price = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
     stock =db.Column(db.Integer, nullable=False)
 
     # Inicializa la clase `Animal`
